// src/utils/types/overview.d.ts
export type OverviewCardKey = 'clients-overview-card' | 'leads-overview-card' | 'users-overview-card' | 'drive-overview-card' | 'meetings-overview-card' | 'onboarding-overview-card' | 'conversion-overview-card' | 'satisfaction-overview-card';

export type OverviewCardsState = Record<OverviewCardKey, boolean>;